Add codes dealing with transported variables
If a variable $\xi$ satifies either
                     \xi_t + u*\xi_x + v*\xi_y = 0,
or
        (\rho*\xi)_t + (\rho*u*\xi)_x + (\rho*v*\xi)_y = 0,
GRP for $\xi$ is solved now.

The solver can deal with an arbitrary number of such variables. Their
initial left states of are given in 'wL.VAR.trans', 'wL.DER.trans' and
'wL.TGT.trans'. Their initial right state are given in the same way in
'wR'.

The results are given in 'out.VAR.trans' and 'out.DER.trans'. The analy-
sis of solving GRPs for such transported variables are given in the
mathematical instruction.