A bug is corrected
In non-trivial, non-sonic cases, the last step is to calculate $\rho_t$.
If the t-axis lies between the contact discontinuity and a shock, the
square of the sound speed is used. This is the variable 'c_square' in
the program. However, we forgot to assign values to that variable. This
is now corrected by adding  codes in line 378 and line 390.